refactor : GameResult 클래스 인스턴스 변수를 Map으로 변경
